app/requirements.txt
Flask==2.2.2
pytest==7.2.0
Flask-AppBuilder==4.1.6
MarkupSafe==2.1.1
WTForms==3.0.1
psycopg2-binary==2.9.5
Bootstrap-Flask==2.1.0
flask-talisman==1.0.0
